Create circuit_graph module and tests
